rpi5_gpio_over_docker_demo/sensors/requirements.txt
flask
opencv-python
gpiozero==2.0
luma.oled